+++ client/src/services/userServces.js
@@ -2,18 +2,54 @@ import api from "../api/axios";
 
 
 // Accessing backend create account api
-export const createAccount = (formData) => {
-    return api.post("/users/auth/create-account", formData)
+export const createAccount = (data) => {
+    return api.post("/users/auth/create-account", data)
 };
 
-// Accessing backend login api
-export const login = (formData) => {
-    return api.post("/users/auth/login", formData)
-};
-
-// Accessing backend verification otp code api
-export const verifyUserCreateAccount = (data) => {
+// Verify verification code for create
+export const verifyCreateAccount = (data) => {
     return api.post("users/auth/create-account/verify",  data);
 }
+
+// Resending verification code for create account
+export const resendCodeTOCreateAccount = (data) => {
+    return api.post("users/auth/create-account/verify/resend",  data);
+}
+
+// Accessing backend login api
+export const login = (data) => {
+    return api.post("/users/auth/login", data)
+};
+
+// Resending verification code for login 
+export const verifyLogin = (data) => {
+    return api.post("/users/auth/login/verify",  data);
+}  
+
+// Resending verification code for login 
+export const resendCodeTOLogin = (data) => {
+    return api.post("users/auth/login/verify/resend",  data);
+}
+
+// Forgot password 
+export const forgotPassword = (data) => {
+    return api.post("users/auth/forgot-password", data);
+}
+
+// Verify verification code for reset password 
+export const verifyResetPassword = (data) => {
+    return api.post("users/auth/reset-password/verify", data);
+}
+
+// Resending verification code for reset password
+export const resendCodeTOResetPassword = (data) => {
+    return api.post("users/auth/reset-password/verify/resend", data);
+}
+
+// Reset password
+export const resetPassword = (data) => {
+    return api.post("users/auth/reset-password", data);
+}
+