+++ server/index.js
@@ -5,6 +5,7 @@ import helmet from "helmet";
 import { connectDB } from "./configs.js";
 import userRouter from "./routes/user.route.js";
 import { globalLimiter } from "./middlewares/security.js";
+import errorHandler from "./middlewares/errorHandler.js";
 dotenv.config();
 
 const app = express();
@@ -26,6 +27,8 @@ app.use('/api/v1', globalLimiter);
 
 // Authentificaton routes api endpoint
 app.use('/api/v1/users/auth', userRouter);
+
+app.use(errorHandler);
 
 try {
     await connectDB();